fix(frontend): the lane tone is one helper, and the render split holds the cap
The three warn-toned lanes share one warnWhenHolding instead of three
ternaries, which is also what keeps TodayLanes under the complexity
ceiling; the accountability cards (failedItem, dsrItem) move to their own
file so render.go stays under the length cap.
